+++ tests/mocks/MockProductionModel.h
@@ -38,6 +38,7 @@ public:
     MOCK_METHOD(model::QualityCheckpoint, getQualityCheckpoint, (uint32_t id), (const, override));
     MOCK_METHOD((std::vector<model::QualityCheckpoint>), getQualityCheckpoints, (), (const, override));
     MOCK_METHOD(model::WorkUnit, getWorkUnit, (), (const, override));
+    MOCK_METHOD(std::string, lastFaultReason, (), (const, override));
 };
 
 }  // namespace app::test